backend/requirements.txt
fastapi
uvicorn
sqlalchemy
requests
beautifulsoup4
python-dotenv
google-genai
pydantic
